new branch for rework of search
